Android-Web-Os
This is a project that aims to build a completely web based OS, that
looks like, feels like and has all the functionality like Android M.
This opens up a new universe of possibilities for websites to be more
lively, functional, and personal ( Goodbye to the days of just 'Good
Looking' static text websites).

To run this Project on your system, you need to install this list of
softwares in this specific order on your system, as by testing using the
good old method of 'trial and error' this order of installing them works
the best for everyone.
1) Node JS
2) NPM
3) Bower
4) Polymer by Google [ recommended to do using Bower only, as it works
the best like that to maintain and to use the Polymer files in day to
day life { no more spaghetti code !! }, for more details visit the
Polymer website.]
5) The clone of the website on your system ( Obviously :P )
6) All the Polymer Element Catalogue's elements downloaded on your
system through Bower [ or you can download the dependencies/libraries
only used in the website { this is mentioned in the website code itself
in the <link></link> tags (ignore the file path) in the <head></head>
tag }, but you will have to install the new dependencies used, later
through Bower or the new features will not run on your system]
7) Apache installed and running on your system ( the software
recommended to accomplish this is WAMP )

Once you have all of this installed (Whew), you can run the website on
your system, and future revisions of it.